Preserve preprocessor-compatible KLI error aliases

+++ programs/libkarm/errno.h
@@ -14,12 +14,12 @@
 #include "include/armonios/abi/errors.h"
 
 /* Historical libkarm spellings retained for source compatibility. */
-#define KLI_NOMEM ((long)ARMONIOS_ERR_NOMEM)
-#define KLI_NOENT ((long)ARMONIOS_ERR_NOENT)
-#define KLI_BADF  ((long)ARMONIOS_ERR_BADF)
-#define KLI_INVAL ((long)ARMONIOS_ERR_INVAL)
-#define KLI_AGAIN ((long)ARMONIOS_ERR_AGAIN)
-#define KLI_PERM  ((long)ARMONIOS_ERR_PERM)
+#define KLI_NOMEM ARMONIOS_ERR_NOMEM
+#define KLI_NOENT ARMONIOS_ERR_NOENT
+#define KLI_BADF  ARMONIOS_ERR_BADF
+#define KLI_INVAL ARMONIOS_ERR_INVAL
+#define KLI_AGAIN ARMONIOS_ERR_AGAIN
+#define KLI_PERM  ARMONIOS_ERR_PERM
 
 static inline int kli_isok(long ret) {
     return ret >= 0;
